Modificaciones de tipado y acceso

+++ SistemaFacturacion/src/electrodomesticos/Televisor.java
@@ -1,16 +1,16 @@
 package electrodomesticos;
 
 public class Televisor extends Electrodomestico {
-    private float tamano;
+    private int tamano;
     private boolean tieneTDT;
 
-    public Televisor(char consumo, String procedencia, float tamano, boolean tieneTDT) {
+    public Televisor(char consumo, String procedencia, int tamano, boolean tieneTDT) {
         super(consumo, procedencia);
         this.tamano = tamano;
         this.tieneTDT = tieneTDT;
     }
 
-    public float getTamano() {
+    public int getTamano() {
         return tamano;
     }
